requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_server]
